requirements.txt
asgiref==3.2.7
Django==3.0.7
django-mysql==3.6.0
djangorestframework==3.11.0
PyMySQL==0.9.3
pytz==2020.1
sqlparse==0.3.1
